test(discovery): move phase-2 coverage tests into the excluded file
My previous commit (3cacff2) added the discovery mock doubles inline in
discovery.rs to cover the timeout branches — but that file is NOT coverage-
excluded, so the ~30 `unimplemented!()` trait stubs counted as 65 new
*uncovered* patch lines and made codecov/patch worse, not better.

Relocate the DiscoveryStubDriver/DiscoveryStubElement doubles and the three
phase-2 tests into page_object_tests.rs, which is already in codecov's ignore
list precisely because its mock stubs never run. find_known_page_objects is
`pub`, so it's called as crate::discovery::find_known_page_objects from there.
discovery.rs is reverted to its clean state.

Net: the 12 real timeout-branch lines in discovery.rs are covered; the stub
boilerplate no longer skews patch coverage. Verified via cargo llvm-cov — zero
uncovered added lines in non-excluded files (the one residual, page_object.rs
:761, is a pre-existing closing brace from f28bd2e, untouched here).
